refactor: phase0 networking
