+++ src/networking.rs
@@ -1,23 +1,12 @@
-use crate::primitives::Epoch;
 use enr;
 pub use multiaddr::Multiaddr;
 use multihash::{Code, Error, Multihash};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
-use ssz_rs::prelude::Bitvector;
 use std::fmt;
-use std::time::Duration;
 use std::{convert::TryFrom, str::FromStr};
 use thiserror::Error;
-pub const ATTESTATION_SUBNET_COUNT: usize = 64;
-pub const GOSSIP_MAX_SIZE: usize = 2usize.pow(20);
-pub const MAX_REQUEST_BLOCKS: usize = 2usize.pow(10);
-pub const MIN_EPOCHS_FOR_BLOCK_REQUESTS: Epoch = 33024;
-pub const MAX_CHUNK_SIZE: usize = 2usize.pow(20);
-pub const TTFB_TIMEOUT: Duration = Duration::from_secs(5);
-pub const RESP_TIMEOUT: Duration = Duration::from_secs(10);
-pub const ATTESTATION_PROPAGATION_SLOT_RANGE: usize = 32;
-pub const MAXIMUM_GOSSIP_CLOCK_DISPARITY: Duration = Duration::from_millis(500);
+
 pub const MAX_INLINE_KEY_LENGTH: usize = 42;
 
 // PeerId reimplemented from rust-libp2p
@@ -135,13 +124,6 @@ pub type Enr = enr::Enr<enr::k256::ecdsa::SigningKey>;
 pub enum MessageDomain {
     InvalidSnappy,
     ValidSnappy,
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub struct MetaData {
-    #[serde(with = "crate::serde::as_string")]
-    pub seq_number: u64,
-    pub attnets: Bitvector<ATTESTATION_SUBNET_COUNT>,
 }
 
 #[cfg(test)]
