+++ src/phase0/mod.rs
@@ -8,6 +8,7 @@ mod epoch_processing;
 mod fork;
 mod genesis;
 mod helpers;
+mod networking;
 mod operations;
 mod presets;
 mod slot_processing;
@@ -22,6 +23,7 @@ pub use epoch_processing::*;
 pub use fork::*;
 pub use genesis::*;
 pub use helpers::*;
+pub use networking::*;
 pub use operations::*;
 pub use presets::Preset;
 pub use slot_processing::*;
